Refactor file parsers, but causes issues with order of execution
