+++ app/jobs/gtfs_import.rb
@@ -1,6 +1,15 @@
 require 'httparty'
 require 'zip'
 
+require_relative "./gtfs_import/zip_file_parser"
+require_relative "./gtfs_import/zip_file_parsers/agency_file_parser"
+require_relative "./gtfs_import/zip_file_parsers/calendars_file_parser"
+require_relative "./gtfs_import/zip_file_parsers/calendar_dates_file_parser"
+require_relative "./gtfs_import/zip_file_parsers/routes_file_parser"
+require_relative "./gtfs_import/zip_file_parsers/stops_file_parser"
+require_relative "./gtfs_import/zip_file_parsers/stop_times_file_parser"
+require_relative "./gtfs_import/zip_file_parsers/trips_file_parser"
+
 class GtfsImport < ApplicationJob
   queue_as :default
 
@@ -65,256 +74,23 @@ class GtfsImport < ApplicationJob
     end
 
     Zip::File.open(@destination_path) do |zip_file|
-      parse_agency(zip_file) # GtfsImport::AgencyImport.new(zip_file).perform
-      parse_calendar(zip_file) # GtfsImport::CalendarImport.new(zip_file).perform
-      parse_calendar_dates(zip_file)
-      parse_routes(zip_file)
-      parse_stops(zip_file)
-      parse_trips(zip_file)
-      parse_stop_times(zip_file) #NOTE: depends on successful completion of #parse_stops and #parse_trips
+      options = {:zip_file => zip_file, :schedule => @schedule, :logger => @logger}
+      AgencyFileParser.new(options).perform
+      CalendarsFileParser.new(options).perform
+      CalendarDatesFileParser.new(options).perform
+      RoutesFileParser.new(options).perform
+      StopsFileParser.new(options).perform
+      TripsFileParser.new(options).perform #NOTE: depends on successful completion of routes
+      StopTimesFileParser.new(options).perform #NOTE: depends on successful completion of stops and trips
     end
+    # ok, so these performances are not executing sequentially for some reason
+    # ... this immediately throws "ActiveRecord::RecordInvalid: Validation failed: Route must exist" because it is trying to parse routes before trips are done.
+    # ... WAT?
+    # ... http://stackoverflow.com/questions/41495844/ruby-how-to-enforce-sequential-execution
   end
 
   def activate
     @logger.info{ "ACTIVATING SCHEDULE" }
     @schedule.activate!
   end
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-  # @param [String] str An exception code (e.g. "1", "2", or "1014")
-  def parse_numeric(str)
-    str.to_i unless str.blank?
-  end
-
-  # @param [String] str A boolean-convertable integer (either "0" or "1")
-  def parse_bool(str)
-    case str; when "0"; false; when "1"; true; end
-  end
-
-  # @param [String] str A latitude or longitude decimal (e.g. "41.29771887088102" or " -72.92673110961914")
-  def parse_decimal(str)
-    str.strip.to_f.round(8)
-  end
-
-  def read_file(zip_file, entry_name)
-    entry = zip_file.entries.find{|entry| entry.name == entry_name }
-    return entry.get_input_stream.read
-  end
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-  # @see https://developers.google.com/transit/gtfs/reference/agency-file
-  def parse_agency(zip_file)
-    @logger.info{ "IMPORTING AGENCIES" }
-    results = read_file(zip_file, "agency.txt")
-    CSV.parse(results, :headers => true) do |row|
-      agency = Agency.where(:schedule_id => @schedule.id, :url => row["agency_url"]).first_or_initialize
-      agency.update!({
-        :guid => row["agency_id"],
-        :name => row["agency_name"] || row[" agency_name"], # handle malformed header name
-        :timezone => row["agency_timezone"],
-        :phone => row["agency_phone"],
-        :lang => row["agency_lang"]
-      })
-    end
-    @logger.info{ "... #{Agency.count}" }
-  end
-
-  # @see https://developers.google.com/transit/gtfs/reference/calendar-file
-  def parse_calendar(zip_file)
-    @logger.info{ "IMPORTING CALENDARS" }
-    results = read_file(zip_file, "calendar.txt")
-    CSV.parse(results, :headers => true) do |row|
-      calendar = Calendar.where(:schedule_id => @schedule.id, :service_guid => row["service_id"]).first_or_initialize
-      calendar.update({
-        :monday => parse_bool(row["monday"]),
-        :tuesday => parse_bool(row["tuesday"]),
-        :wednesday => parse_bool(row["wednesday"]),
-        :thursday => parse_bool(row["thursday"]),
-        :friday => parse_bool(row["friday"]),
-        :saturday => parse_bool(row["saturday"]),
-        :sunday => parse_bool(row["sunday"]),
-        :start_date => row["start_date"].to_date,
-        :end_date => row["end_date"].to_date
-      })
-    end
-    @logger.info{ "... #{Calendar.count}" }
-  end
-
-  # @see https://developers.google.com/transit/gtfs/reference/calendar_dates-file
-  def parse_calendar_dates(zip_file)
-    @logger.info{ "IMPORTING CALENDAR DATES" }
-    results = read_file(zip_file, "calendar_dates.txt")
-    CSV.parse(results, :headers => true) do |row|
-      calendar_date = CalendarDate.where({
-        :schedule_id => @schedule.id,
-        :service_guid => row["service_id"],
-        :exception_date => row["date"].to_date
-      }).first_or_initialize
-      calendar_date.update!(:exception_code => parse_numeric(row["exception_type"]))
-    end
-    @logger.info{ "... #{CalendarDate.count}" }
-  end
-
-  # @see https://developers.google.com/transit/gtfs/reference/routes-file
-  def parse_routes(zip_file)
-    @logger.info{ "IMPORTING ROUTES" }
-    results = read_file(zip_file, "routes.txt")
-    CSV.parse(results, :headers => true) do |row|
-      route = Route.where(:schedule_id => @schedule.id, :guid => row["route_id"]).first_or_initialize
-      route.update!({
-        :agency_guid => row["agency_id"],
-        :short_name => row["route_short_name"],
-        :long_name => row["route_long_name"],
-        :description => row["route_desc"],
-        :code => parse_numeric(row["route_type"]),
-        :url => row["url"],
-        :color => row["color"],
-        :text_color => row["text_color"],
-      })
-    end
-    @logger.info{ "... #{Route.count}" }
-  end
-
-  # @see https://developers.google.com/transit/gtfs/reference/stops-file
-  def parse_stops(zip_file)
-    @logger.info{ "IMPORTING STOPS" }
-    results = read_file(zip_file, "stops.txt")
-    CSV.parse(results, :headers => true) do |row|
-      stop = Stop.where(:schedule_id => @schedule.id, :guid => row["stop_id"]).first_or_initialize
-      stop.update!({
-        :short_name => row["stop_code"],
-        :name => row["stop_name"],
-        :description => row["stop_desc"],
-        :latitude => parse_decimal(row["stop_lat"]),
-        :longitude => parse_decimal(row["stop_lon"]),
-        :zone_guid => row["zone_id"],
-        :url => row["stop_url"],
-        :location_code => parse_numeric(row["location_type"]),
-        :parent_guid => row["parent_station"],
-        :timezone => row["stop_timezone"],
-        :wheelchair_code => parse_numeric(row["wheelchair_boarding"])
-      })
-    end
-    @logger.info{ "... #{Stop.count}" }
-  end
-
-  # @see https://developers.google.com/transit/gtfs/reference/trips-file
-  def parse_trips(zip_file)
-    @logger.info{ "IMPORTING TRIPS" }
-    results = read_file(zip_file, "trips.txt")
-    CSV.parse(results, :headers => true) do |row|
-      trip = Trip.where(:schedule_id => @schedule.id, :guid => row["trip_id"]).first_or_initialize
-      trip.update!({
-        :route_guid => row["route_id"],
-        :service_guid => row["service_id"],
-        :headsign => row["trip_headsign"],
-        :short_name => row["trip_short_name"],
-        :direction_code => parse_numeric(row["direction_id"]),
-        :block_guid => row["block_id"],
-        :shape_guid => row["shape_id"],
-        :wheelchair_code => parse_numeric(row["wheelchair_accessible"]),
-        :bicycle_code => parse_numeric(row["bikes_allowed"])
-      })
-    end
-    @logger.info{ "... #{CalendarDate.count}" }
-  end
-
-  # @see https://developers.google.com/transit/gtfs/reference/stop_times-file
-  def parse_stop_times(zip_file)
-    @logger.info{ "IMPORTING STOP TIMES" }
-    results = read_file(zip_file, "stop_times.txt")
-    CSV.parse(results, :headers => true) do |row|
-      stop_time = StopTime.where({
-        :schedule_id => @schedule.id,
-        :trip_guid => row["trip_id"],
-        :stop_guid => row["stop_id"],
-      }).first_or_initialize
-
-      # tracking a potential issue with the gtfs data...
-      # there are sequential stop_times which share a composite key but differ in sequence.
-      # @see https://gist.github.com/s2t2/0d2929e0ecaba85823e1314935e7941e
-      # consider importing all records and adding stop_sequence to the composite key.
-      if stop_time.persisted?
-        raise UnexpectedStopTime.new(row.to_h) unless stop_time.stop_sequence + 1 == row.to_h['stop_sequence'].to_i
-        @logger.info{ "... CONSOLIDATING STOP TIMES #{stop_time.trip_guid}-#{stop_time.stop_guid} (#{stop_time.stop_sequence} vs #{row['stop_sequence']})" }
-        stop_time.update!(:departure_time => row["departure_time"], :stop_sequence => row["stop_sequence"].to_i)
-      else
-        stop_time.update!({
-          :stop_sequence => row["stop_sequence"].to_i,
-          :arrival_time => row["arrival_time"],
-          :departure_time => row["departure_time"],
-          :headsign => row["stop_headsign"],
-          :pickup_code => parse_numeric(row["pickup_type"]),
-          :dropoff_code => parse_numeric(row["drop_off_type"]),
-          :distance => row["shape_dist_traveled"],
-          :code => parse_numeric(row["timepoint"])
-        })
-      end
-    end
-    @logger.info{ "... #{StopTime.count}" }
-  end
-  class UnexpectedStopTime < StandardError ; end
-
-
-
-
-
-
-
-
-
-
-
-
 end
